Move organism components with their stories into seperate directories
