+++ .vscode/settings.json
@@ -28,7 +28,9 @@
         "02_molecules": "App",
         "MenuDesktop": "App",
 		"MenuMobile": "App",
-		"03_organisms": "Components",
+        "03_organisms": "Components",
+        "Footer": "Components",
+        "Header": "Components",
 		"04_templates": "Layout",
 		"05_pages": "Log",
 		"06_posts": "Docs",
